Add files via upload
In this program, we have the isprime() function that checks whether a number is prime or not. The main program repeatedly asks the user to input a number between 2 and 100. If the number is out of range, it prompts the user to try again or exit. Once a valid input is received, it calls the isprime() function and prints whether the number is prime or not.

Please note that the program assumes valid inputs of integer values. If the user enters non-integer values, it may result in unexpected behavior.